Creating main.cc
Okay so this is going to be our main file that everyone's code is going to eventually go into. I think we should try the different branch system for our three main stuffs (Preet, Denise, and mine) but we'll see. 
We shouldn't need any header files since we're going to use the STL heap and hash table.